+++ backend/canisters/user/impl/src/updates/send_message.rs
@@ -1,4 +1,5 @@
 use crate::crypto::process_transaction_without_caller_check;
+use crate::guards::caller_is_local_user_index;
 use crate::guards::caller_is_owner;
 use crate::model::pin_number::VerifyPinError;
 use crate::timer_job_types::{DeleteFileReferencesJob, MarkP2PSwapExpiredJob, NotifyEscrowCanisterOfDepositJob};
@@ -7,16 +8,28 @@ use crate::{mutate_state, read_state, run_regular_jobs, Data, RuntimeState, Time
 use candid::Principal;
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
+use chat_events::EditMessageArgs;
+use chat_events::EditMessageResult;
 use chat_events::{MessageContentInternal, PushMessageArgs, Reader, ReplyContextInternal, ValidateNewMessageContentResult};
 use constants::{MEMO_MESSAGE, OPENCHAT_BOT_USER_ID};
+use event_store_producer_cdk_runtime::CdkRuntime;
 use rand::Rng;
+use types::BotCaller;
+use types::BotPermissions;
+use types::DirectMessageNotification;
+use types::EventIndex;
+use types::Notification;
 use types::{
     BlobReference, CanisterId, Chat, ChatId, CompletedCryptoTransaction, ContentValidationError, CryptoTransaction,
     EventWrapper, Message, MessageContent, MessageContentInitial, MessageId, MessageIndex, P2PSwapLocation, ReplyContext,
     TimestampMillis, UserId, UserType,
 };
+use user_canister::c2c_bot_send_message;
 use user_canister::send_message_v2::{Response::*, *};
 use user_canister::{C2CReplyContext, SendMessageArgs, SendMessagesArgs, UserCanisterEvent};
+
+use super::c2c_send_messages::handle_message_impl;
+use super::c2c_send_messages::HandleMessageArgs;
 
 #[update(guard = "caller_is_owner", msgpack = true)]
 #[trace]
@@ -28,7 +41,7 @@ async fn send_message_v2(args: Args) -> Response {
         now,
         local_user_index_canister_id,
         maybe_recipient_type,
-    } = match read_state(|state| prepare(&args, state)) {
+    } = match read_state(|state| prepare(&args, false, state)) {
         Ok(ok) => ok,
         Err(response) => return *response,
     };
@@ -161,6 +174,164 @@ async fn send_message_v2(args: Args) -> Response {
     })
 }
 
+#[update(guard = "caller_is_local_user_index", msgpack = true)]
+#[trace]
+fn c2c_bot_send_message(args: c2c_bot_send_message::Args) -> c2c_bot_send_message::Response {
+    run_regular_jobs();
+
+    mutate_state(|state| {
+        let finalised = args.finalised;
+        let bot_id = args.bot_id;
+        let bot_name = args.bot_name.clone();
+        let bot_caller = BotCaller {
+            bot: args.bot_id,
+            initiator: args.initiator.clone(),
+        };
+
+        let args: Args = args.into();
+        let message_content: MessageContent = args.content.clone().into();
+
+        if !state.data.is_bot_permitted(
+            &bot_id,
+            &bot_caller.initiator,
+            BotPermissions::from_message_permission((&args.content).into()),
+        ) {
+            return c2c_bot_send_message::Response::NotAuthorized;
+        }
+
+        let result = match prepare(&args, true, state) {
+            Ok(ok) => ok,
+            Err(response) => return (*response).into(),
+        };
+
+        let now = result.now;
+
+        let content =
+            match MessageContentInternal::validate_new_message(args.content, true, UserType::BotV2, args.forwarding, now) {
+                ValidateNewMessageContentResult::Success(content) => content,
+                ValidateNewMessageContentResult::SuccessP2PSwap(_)
+                | ValidateNewMessageContentResult::SuccessCrypto(_)
+                | ValidateNewMessageContentResult::SuccessPrize(_) => unreachable!(),
+                ValidateNewMessageContentResult::Error(error) => {
+                    let response = match error {
+                        ContentValidationError::Empty => MessageEmpty,
+                        ContentValidationError::TextTooLong(max_length) => TextTooLong(max_length),
+                        ContentValidationError::InvalidPoll(reason) => InvalidPoll(reason),
+                        ContentValidationError::TransferCannotBeZero => TransferCannotBeZero,
+                        ContentValidationError::InvalidTypeForForwarding => {
+                            InvalidRequest("Cannot forward this type of message".to_string())
+                        }
+                        ContentValidationError::TransferMustBePending | ContentValidationError::PrizeEndDateInThePast => {
+                            unreachable!()
+                        }
+                        ContentValidationError::Unauthorized => {
+                            InvalidRequest("User unauthorized to send messages of this type".to_string())
+                        }
+                    };
+
+                    return response.into();
+                }
+            };
+
+        // Check if a message with the same id already exists
+        if let Some(chat) = state.data.direct_chats.get_mut(&bot_id.into()) {
+            if let Some((message, _)) =
+                chat.events
+                    .message_internal(EventIndex::default(), args.thread_root_message_index, args.message_id.into())
+            {
+                // If the message id of a bot message matches an existing unfinalised bot message
+                // then edit this message instead of pushing a new one
+                if let Some(bot_message) = message.bot_context {
+                    if bot_caller.bot == message.sender
+                        && bot_caller.initiator.user() == bot_message.command.as_ref().map(|c| c.initiator)
+                        && bot_caller.initiator.command() == bot_message.command.as_ref()
+                        && !bot_message.finalised
+                    {
+                        let edit_message_args = EditMessageArgs {
+                            sender: bot_caller.bot,
+                            min_visible_event_index: EventIndex::default(),
+                            thread_root_message_index: args.thread_root_message_index,
+                            message_id: args.message_id,
+                            content,
+                            block_level_markdown: Some(args.block_level_markdown),
+                            finalise_bot_message: finalised,
+                            now,
+                        };
+
+                        let EditMessageResult::Success(message_index, event) =
+                            chat.events.edit_message::<CdkRuntime>(edit_message_args, None)
+                        else {
+                            // Shouldn't happen
+                            return c2c_bot_send_message::Response::NotAuthorized;
+                        };
+
+                        if finalised && !chat.notifications_muted.value {
+                            let notification = Notification::DirectMessage(DirectMessageNotification {
+                                sender: bot_id,
+                                thread_root_message_index: args.thread_root_message_index,
+                                message_index,
+                                event_index: event.index,
+                                sender_name: bot_name,
+                                sender_display_name: None,
+                                message_type: message_content.message_type(),
+                                message_text: message_content.notification_text(&[], &[]),
+                                image_url: message_content.notification_image_url(),
+                                sender_avatar_id: None,
+                                crypto_transfer: message_content.notification_crypto_transfer_details(&[]),
+                            });
+                            let recipient = state.env.canister_id().into();
+
+                            state.push_notification(Some(bot_id), recipient, notification);
+                        }
+
+                        return c2c_bot_send_message::Response::Success(SuccessResult {
+                            chat_id: bot_id.into(),
+                            event_index: event.index,
+                            message_index,
+                            expires_at: event.expires_at,
+                            timestamp: now,
+                        });
+                    }
+                }
+
+                return c2c_bot_send_message::Response::MessageAlreadyFinalised;
+            }
+        }
+
+        let event_wrapper = handle_message_impl(
+            HandleMessageArgs {
+                sender: bot_id,
+                thread_root_message_id: None,
+                message_id: Some(args.message_id),
+                sender_message_index: None,
+                sender_name: bot_name,
+                sender_display_name: None,
+                content,
+                replies_to: None,
+                forwarding: false,
+                sender_user_type: UserType::BotV2,
+                sender_avatar_id: None,
+                push_message_sent_event: true,
+                mentioned: Vec::new(),
+                mute_notification: !finalised,
+                block_level_markdown: args.block_level_markdown,
+                now,
+            },
+            Some(bot_caller),
+            finalised,
+            state,
+        );
+
+        c2c_bot_send_message::Response::Success(SuccessResult {
+            chat_id: bot_id.into(),
+            event_index: event_wrapper.index,
+            message_index: event_wrapper.event.message_index,
+            expires_at: event_wrapper.expires_at,
+            timestamp: now,
+        })
+    })
+}
+
 #[derive(Copy, Clone)]
 enum RecipientType {
     _Self,
@@ -193,7 +364,7 @@ struct PrepareOk {
     maybe_recipient_type: Option<RecipientType>,
 }
 
-fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareOk, Box<Response>> {
+fn prepare(args: &Args, is_v2_bot: bool, state: &RuntimeState) -> Result<PrepareOk, Box<Response>> {
     if state.data.suspended.value {
         return Err(Box::new(UserSuspended));
     }
@@ -212,7 +383,7 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareOk, Box<Response>
     let maybe_recipient_type = if let Some(chat) = state.data.direct_chats.get(&args.recipient.into()) {
         if chat
             .events
-            .contains_message_id(args.thread_root_message_index, args.message_id)
+            .message_already_finalised(args.thread_root_message_index, args.message_id, is_v2_bot)
         {
             return Err(Box::new(DuplicateMessageId));
         }